add user info crud, model, controller, seeder, migration

+++ app/Http/Controllers/Admin/PostController.php
@@ -5,14 +5,16 @@ namespace App\Http\Controllers\Admin;
 use App\Model\Post;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
     protected $validationParams = [
         'title' => 'required|max:255',
         'content' => 'required',
+        'category_id' => 'exists:App\Model\Category,id'
     ];
+
 
     /**
      * Display a listing of the resource.
@@ -21,9 +23,24 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::paginate(20);
+        $posts = Post::orderBy('created_at', 'desc')->paginate(20);
+
         return view('admin.posts.index', compact('posts'));
     }
+
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function indexUser()
+    {
+        $posts = Post::where('user_id', Auth::user()->id)->orderBy('created_at', 'desc')->paginate(20);
+
+        return view('admin.posts.index', ['posts' => $posts]);
+    }
+
 
     /**
      * Show the form for creating a new resource.
@@ -32,8 +49,11 @@ class PostController extends Controller
      */
     public function create()
     {
+        // $categories = Category::all();
+
         return view('admin.posts.create');
     }
+
 
     /**
      * Store a newly created resource in storage.
@@ -43,28 +63,19 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $data = $request->all();
+        $data['user_id'] = Auth::user()->id;
+
         $request->validate($this->validationParams);
 
-        $data = $request->all();
+        $post = new Post();
+        $post->fill($data);
+        $post->slug = $post->createSlug($data['title']);
+        $post->save();
 
-        $slug = Str::slug($data['title'], '-');
-        $postPresente = Post::where('slug', $slug)->first();
-
-        $counter = 0;
-        while ($postPresente) {
-            $slug = $slug . '-' . $counter;
-            $postPresente = Post::where('slug', $slug)->first();
-            $counter++;
-        }
-
-        $newPost = new Post();
-
-        $newPost->fill($data);
-        $newPost->slug = $slug;
-        $newPost->save();
-
-        return redirect()->route('admin.posts.show', ['post' => $newPost]);
+        return redirect()->route('admin.posts.show', $post->slug);
     }
+
 
     /**
      * Display the specified resource.
@@ -74,8 +85,9 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        return view('admin.posts.show', compact('post'));
+        return view('admin.posts.show', ['post' => $post]);
     }
+
 
     /**
      * Show the form for editing the specified resource.
@@ -85,9 +97,22 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
-        return view('admin.posts.edit', compact('post'));
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
+
+        // $categories = Category::all();
+
+        $data = [
+            'post' => $post,
+            // 'categories' => $categories
+        ];
+
+        return view('admin.posts.edit', $data);
     }
 
+
+    
     /**
      * Update the specified resource in storage.
      *
@@ -97,15 +122,31 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        $data = $request->all();
+
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
 
         $request->validate($this->validationParams);
 
-        $data = $request->all();
-        $post->update($data);
+        if ($data['title'] != $post->title) {
+            $post->title = $data['title'];
+            $post->slug = $post->createSlug($data['title']);
+        }
+        if ($data['content'] != $post->content) {
+            $post->content = $data['content'];
+        }
+        if ($data['category_id'] != $post->category_id) {
+            $post->category_id = $data['category_id'];
+        }
 
-        return redirect()->route('admin.posts.show', $post->id);
+        $post->update();
+
+        return redirect()->route('admin.posts.show', $post);
     }
 
+    
     /**
      * Remove the specified resource from storage.
      *
@@ -114,6 +155,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if (Auth::user()->id != $post->user_id) {
+            abort('403');
+        }
+
         $post->delete();
 
         return redirect()
